kubernetes/health-check-service/requirements.txt
requests
kafka-python
flask